Added "multiline" flag to errorMatch regex
When a build fails, depending on the compiler's output, the error message could span multiple lines. If the location of the error within the source code is not specified on the first line, the build:error-match function cannot jump to the right location. By enabling multiline regexes, one could match against the whole output.

I stumbled upon this issue while trying to add a build configuration to a Nim project.